Source: FINALLY remove copy_board
bench: 2928707

+++ Source/structs.h
@@ -70,9 +70,9 @@ typedef struct position {
   uint64_t bitboards[12];
   uint64_t occupancies[3];
   hash_keys_t hash_keys;
-  uint32_t ply;
-  uint32_t seldepth;
-  uint32_t fifty;
+  uint8_t ply;
+  uint8_t seldepth;
+  uint8_t fifty;
   uint8_t mailbox[64];
   uint8_t side;
   uint8_t enpassant;
